Big change: html -> markdown
Moved the book from being written in raw html to being written in
markdown.  This was mostly automated with pandoc, but required
some finagling to get things looking just right.  There are still
a few minor unresolved things, but I'll knock those out ere long.
The main reason for doing this is to ease having a PDF version :)